bot/requirements.txt
aiogram==3.16.0
sqlalchemy==2.0.41
asyncpg==0.30.0
python-dotenv==1.0.1